Move more interfaces to correct directories
