+++ frontend/interfaces/software.ts
@@ -10,6 +10,19 @@ export default PropTypes.shape({
   vulnerabilities: PropTypes.arrayOf(vulnerabilityInterface),
 });
 
+export interface ISoftwareResponse {
+  counts_updated_at: string;
+  software: ISoftware[];
+}
+
+export interface ISoftwareCountResponse {
+  count: number;
+}
+
+export interface IGetSoftwareByIdResponse {
+  software: ISoftware;
+}
+
 export interface ISoftware {
   hosts_count?: number;
   id: number;
